Replace all misuses of "int" and "long" with "size_t" in src/ptr-array.c

+++ src/ptr-array.h
@@ -5,8 +5,8 @@
 
 typedef struct {
     void **ptrs;
-    long alloc;
-    long count;
+    size_t alloc;
+    size_t count;
 } PointerArray;
 
 #define PTR_ARRAY_INIT { \
@@ -19,12 +19,12 @@ typedef void (*free_func)(void *ptr);
 #define FREE_FUNC(f) (free_func)f
 
 void ptr_array_add(PointerArray *array, void *ptr);
-void ptr_array_insert(PointerArray *array, void *ptr, long pos);
+void ptr_array_insert(PointerArray *array, void *ptr, size_t pos);
 void ptr_array_free_cb(PointerArray *array, free_func free_ptr);
 void ptr_array_remove(PointerArray *array, void *ptr);
-void *ptr_array_remove_idx(PointerArray *array, long pos);
-long ptr_array_idx(PointerArray *array, void *ptr);
-void *ptr_array_rel(PointerArray *array, void *ptr, long offset);
+void *ptr_array_remove_idx(PointerArray *array, size_t pos);
+size_t ptr_array_idx(PointerArray *array, void *ptr);
+void *ptr_array_rel(PointerArray *array, void *ptr, size_t offset);
 
 static inline void *ptr_array_next(PointerArray *array, void *ptr)
 {
